Integrate SPDX license header check with tox
