+++ api/app/signals/apps/api/v1/serializers/empty.py
@@ -1,3 +1,5 @@
+# SPDX-License-Identifier: MPL-2.0
+# Copyright (C) 2020 - 2021 Vereniging van Nederlandse Gemeenten
 from rest_framework import serializers
