+++ api/app/tox.ini
@@ -1,5 +1,5 @@
 [tox]
-envlist = pytest,flake8,isort
+envlist = pytest,flake8,isort,spdx
 skipsdist = True
 toxworkdir = /tmp/
 
@@ -17,6 +17,7 @@ commands =
     pytest: pytest -n 4 --cov=signals --cov-report=term --cov-report=html:{toxworkdir}/test/htmlcov --no-cov-on-fail {posargs:} --tb=short
     flake8: flake8 signals tests
     isort: isort --recursive --diff --check-only signals tests
+    spdx: python check_spdx.py .
 
 [pytest]
 DJANGO_SETTINGS_MODULE = signals.settings.testing
